feat: remove badge in favour of always showing the civic button + rename identity related components
