+++ .storybook/main.js
@@ -31,7 +31,7 @@ module.exports = {
 		config.resolve.alias = {
 			...config.resolve?.alias,
 			"@components": path.resolve(__dirname, "../src/components"),
-			"@types": path.resolve(__dirname, "../src/types"),
+			"@credix_types": path.resolve(__dirname, "../src/types"),
 			"@": [path.resolve(__dirname, "../src/"), path.resolve(__dirname, "../")],
 		};
