+++ jest.config.js
@@ -14,7 +14,7 @@ const customJestConfig = {
 	testEnvironment: "jest-environment-jsdom",
 	moduleNameMapper: {
 		"^@components/(.*)": "<rootDir>/src/components/$1",
-		"^@types/(.*)": "<rootDir>/src/types/$1",
+		"^@credix_types/(.*)": "<rootDir>/src/types/$1",
 	},
 	modulePathIgnorePatterns: ["__mocks__", ".yalc"],
 };
