+++ src/components/CivicButton.tsx
@@ -1,0 +1,17 @@
+import { useAnchorWallet } from "@solana/wallet-adapter-react";
+import { IdentityButton, useGateway } from "@civic/solana-gateway-react";
+
+export const CivicButton = () => {
+	const wallet = useAnchorWallet();
+	const { requestGatewayToken } = useGateway();
+
+	if (!wallet?.publicKey) {
+		return null;
+	}
+
+	return (
+		<div className="identity-button">
+			<IdentityButton onClick={requestGatewayToken} />
+		</div>
+	);
+};
